% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking_fun.R
\name{blocking_fun}
\alias{blocking_fun}
\title{Create blocking hashes for address matching}
\usage{
blocking_fun(x, address = "input", not_gnaf = FALSE)
}
\arguments{
\item{x}{data.table containing address data}

\item{address}{Character string specifying the address column name}

\item{not_gnaf}{Logical indicating if data is not from G-NAF}
}
\value{
data.table with blocking columns added
}
\description{
Create blocking hashes for address matching
}
