% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check}
\alias{check}
\title{Function to check environment variable
local function}
\usage{
check(var)
}
\arguments{
\item{var}{env. var to check}
}
\description{
Function to check environment variable
local function
}
\examples{
\dontrun{
check("foobar")
}
}
