% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{gc_find}
\alias{gc_find}
\title{Find geocode(s)}
\usage{
gc_find(x, collection = "geocode", full_url, as_sf = TRUE)
}
\arguments{
\item{x}{keys and values to return geometries with}

\item{collection}{set collection from parameter default \code{geocode}.}

\item{full_url}{ability to use gc_find on any mongodb}

\item{as_sf}{return the results as an sf object.}
}
\value{
coordinate geospatial R objects
}
\description{
The main function to find geocodes, using standard MongoDB queries,
with focus on particularly on returing "geometry" documents'
coordinates as useful objects (thinking about sfcs for instance).
}
