% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gc_import}
\alias{gc_import}
\title{Import geocode}
\usage{
gc_import(uri, collection = "geocode", local = TRUE,
  index = "2dsphere")
}
\arguments{
\item{uri}{file to read or remote, flick \code{local} param}

\item{collection}{set collection from parameter.}

\item{local}{assume \code{uri} being local}

\item{index}{a MongoDB compatible index: either 2d or 2dsphere}
}
\value{
number of rows inserted
}
\description{
For now, support geojson file as that is what mongolite package tries to support.
Shape and others later which would also mean, keeping the package deps minimal.
For details see data-raw example of this package how geojson objects are imported.
}
