% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gc_import_sf}
\alias{gc_import_sf}
\title{Import \code{sf} objects into Mongodb}
\usage{
gc_import_sf(x, collection = "geocode", index = "2dsphere",
  mongo_url = "mongodb://localhost:27017", silent = FALSE)
}
\arguments{
\item{x}{geojson file to read or remote, flick \code{local} param.}

\item{collection}{set collection from parameter default \code{geocode}.}

\item{index}{a MongoDB compatible index: either 2d or 2dsphere.}

\item{mongo_url}{custom mongourl to use with \code{collection},
default value is \code{'mongodb://localhost:27017'} used with \code{gc_setup}.}

\item{silent}{show messages, default is \code{FALSE}}
}
\value{
number of rows inserted
}
\description{
TODO: function details
}
