% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{gc_prep_street_db}
\alias{gc_prep_street_db}
\title{Prepare Streets Database}
\usage{
gc_prep_street_db(
  data,
  path = gc_cache_path(),
  year = 2022,
  save_edge = TRUE,
  save_face = TRUE,
  save_cens = FALSE,
  refresh = FALSE
)
}
\arguments{
\item{data}{A data frame containing columns \code{state_code} and either
\code{county_code} or \code{zip_code}, such as the output of  \code{\link[=gc_address]{gc_address()}}.}

\item{path}{The folder where the database will be stored. See
\code{\link[=gc_cache_path]{gc_cache_path()}} and the README for more information.}

\item{year}{The year to download data for.}

\item{save_edge}{Whether to save road edge shapefile information. Needed to
use \code{\link[=gc_code_pt]{gc_code_pt()}}.}

\item{save_face}{Whether to save shapefile face information. Needed to
use \code{\link[=gc_code_block]{gc_code_block()}}.}

\item{save_cens}{Whether to save additional Census geographies (school
districts, NECTAs, etc.) for matching with \code{\link[=gc_code_block]{gc_code_block()}}. Only applies
if \code{save_face=TRUE}.}

\item{refresh}{If \code{TRUE}, force re-downloading of the Census data.}
}
\value{
The data, invisibly.
}
\description{
Download and format address ranges from the Census TIGER/Line API, and then
store in Arrow format. One of \code{county} or \code{zip} must be provided.
}
\examples{
\dontest{
# may take longer than 5 seconds
gc_prep_street_db(data.frame(state_code = "37", county_code = "055"))
}

}
