% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gc_query}
\alias{gc_query}
\title{Generate JSON MongoDB spatial query}
\usage{
gc_query(...)
}
\arguments{
\item{...}{different params. Requires query type which must be
one of \code{geoIntersects}, \code{geoWithin}, \code{near} or \code{nearSphere}.}
}
\value{
JSON valid MongoDB query
}
\description{
See
https://docs.mongodb.com/manual/reference/operator/query-geospatial/
}
