% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{gc_setup}
\alias{gc_setup}
\title{Setup mongodb}
\usage{
gc_setup(db)
}
\arguments{
\item{db}{set database from parameter.}
}
\value{
mongolite::mongo valid URI
}
\description{
Set an R environment variable with database details. This is done so that
username and password of the database is kept away from your code base
deliberaltey.
}
\details{
The functions checks and gives feedback on a valid URL to be used with
\code{mongolite::mongo}'s url parameter. To do this it needs:
\itemize{
\item MONGODB_HOST
\item MONGODB_PORT
\item MONGODB_USER
\item MONGODB_PASS
\item (optional) MONGODB_DB
\item (optional) MONGODB_COLL - for default connection
}
}
\examples{
{
host = Sys.getenv("MONGODB_HOST")
Sys.setenv(MONGODB_HOST="localhost")
gc_setup()
Sys.setenv(MONGODB_HOST=host)
}
}
