% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gc_write_to_mongo}
\alias{gc_write_to_mongo}
\title{Internal write helper function}
\usage{
gc_write_to_mongo(sf.df, con)
}
\arguments{
\item{sf.df}{dataframe to iterate and write to \code{con}}

\item{con}{a mongodb connetion initialized with \code{mongolite::mongo}

assemble a coordinates list from current read_json
like \code{[[[lon,lat], [lon, lat]]]}
unboxed properties}
}
\description{
Internal write helper function
}
