% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode using census geocoder}
\usage{
geocode(data, benchmark = "Current", vintage = "Current")
}
\arguments{
\item{data}{Data set to be geocoded.}

\item{benchmark}{Benchmark}

\item{vintage}{Vintage}
}
\description{
Use the census geocoder to lookup lat/long and census geographies.
}
