% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder-package.R
\docType{package}
\name{geocoder-package}
\alias{geocoder}
\alias{geocoder-package}
\title{geocoder: Efficiently Geocode Addresses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alarm-redist/geocoder}
  \item \url{http://alarm-redist.org/geocoder/}
  \item Report bugs at \url{https://github.com/alarm-redist/geocoder/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu}

Authors:
\itemize{
  \item Cory McCartan \email{cmccartan@g.harvard.edu}
}

}
\keyword{internal}
