% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocoding}
\alias{geocoding}
\title{Geocoding a list of address using Nominatim, US Census, manual input and etc.}
\usage{
geocoding(address_list, threshold = 0.5, max_attempt = 3)
}
\arguments{
\item{address_list}{a list of address that should be geocoded}

\item{threshold}{importance threshold to identify invalid geocoded observation}

\item{max_attempt}{max attempt to fetch the data from US census API}
}
\value{
address, lat, lng, city, county, zipcode, status, importance
}
\description{
Geocoding a list of address using Nominatim, US Census, manual input and etc.
}
