% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{geocoding_census}
\alias{geocoding_census}
\title{Geocoding data using United States Census}
\usage{
geocoding_census(address)
}
\arguments{
\item{address}{address is free-form but it is best form your address like {street}, {city}, {county}, {state}, {postalcode}}
}
\value{
address, lat, lng, city, county, zipcode, status, importance
}
\description{
documentation https://geocoding.geo.census.gov/geocoder/
}
