% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{geocoding_census_reverse}
\alias{geocoding_census_reverse}
\title{Reverse geocoding with United States Census}
\usage{
geocoding_census_reverse(lat, lng)
}
\arguments{
\item{lat}{lat}

\item{lng}{lng}
}
\value{
raw request content
}
\description{
Reverse geocoding with United States Census
}
