% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominatim.R
\name{geocoding_nominatim_reverse}
\alias{geocoding_nominatim_reverse}
\title{Reverse geocoding with Nominatim}
\usage{
geocoding_nominatim_reverse(lat, lng)
}
\arguments{
\item{lat}{lat}

\item{lng}{lng}
}
\value{
address, lat, lng, city, county, zipcode, status, importance
}
\description{
Reverse geocoding with Nominatim
}
