% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominatim.R
\name{geocoding_nominatim_safely_cross_street}
\alias{geocoding_nominatim_safely_cross_street}
\title{Support function for \code{geocoding_nominatim} when there address failed to geocode but contains cross road}
\usage{
geocoding_nominatim_safely_cross_street(address, threshold)
}
\arguments{
\item{address}{address is free-form but it is best form your address like {street}, {city}, {county}, {state}, {postalcode}}

\item{threshold}{importance threshold to identify invalid geocoded observation}
}
\value{
address, lat, lng, city, county, zipcode, status, importance
}
\description{
Support function for \code{geocoding_nominatim} when there address failed to geocode but contains cross road
}
