% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{get_census}
\alias{get_census}
\title{Get census data via `censusapi`}
\usage{
get_census(vars = "B01001_001E", geo,
  key = "0a05d377432e0e7c88dd138d65f79082a5c49eec")
}
\arguments{
\item{vars}{Variables.}

\item{geo}{Geography.}

\item{key}{Census API key.}
}
\description{
Gets census data with a modified geography parameter
}
