% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winkler_funs.R
\name{jw_split_words}
\alias{jw_split_words}
\title{Jaro-Winkler similarity for word-split strings}
\usage{
jw_split_words(vec_1, vec_2, regex = "[^A-Z]")
}
\arguments{
\item{vec_1}{Character vector of first strings}

\item{vec_2}{Character vector of second strings}

\item{regex}{Regular expression for splitting (default "\link{^A-Z}")}
}
\value{
Numeric vector of Jaro-Winkler similarity scores
}
\description{
Jaro-Winkler similarity for word-split strings
}
