% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winkler_funs.R
\name{jwink}
\alias{jwink}
\title{Jaro-Winkler string similarity with word matching}
\usage{
jwink(
  vec_1 = c("kyle john", "k yle j"),
  vec_2 = c("kyle j", "kyle j"),
  type = "1",
  allow_contraction = TRUE
)
}
\arguments{
\item{vec_1}{Character vector of first strings}

\item{vec_2}{Character vector of second strings}

\item{type}{Character indicating type of matching (1, 2, or 3)}

\item{allow_contraction}{Logical allowing word contractions}
}
\value{
Numeric vector of similarity scores
}
\description{
Jaro-Winkler string similarity with word matching
}
