% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_address.R
\name{lookup_address}
\alias{lookup_address}
\title{UPDATE}
\usage{
lookup_address(
  x = c("8 GYMPIE ROAD TIN CAN BAY 4580", "190 MUSGRAVE ROAD RED HILL 4000"),
  deduplicate = TRUE,
  test = FALSE,
  block_4 = FALSE,
  add_vars = c("number_first", "number_last", "street_name", "street_type",
    "street_suffix", "locality_name", "postcode")
)
}
\value{
Returns Geocode Output
}
\description{
UPDATE.
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
