% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_gnaf.R
\name{source_gnaf}
\alias{source_gnaf}
\title{UPDATE}
\usage{
source_gnaf(
  gnaf_core_path = "c:/temp/gnaf.qld.csv",
  states = c("QLD"),
  verbose = TRUE
)
}
\arguments{
\item{gnaf_core_path}{A character path to G-NAF Core.}

\item{states}{A character vector of abbreviated Australian State. Default is \code{"QLD"}.}

\item{verbose}{A logical argument to determine if the function should be verbose or not. Default is \code{TRUE}.}
}
\value{
Returns G-NAF Core wrangled in a way suitable for matching to.
}
\description{
UPDATE.
}
\author{
Kyle Haynes, \email{kyle@kylehaynes.com.au}.
}
