% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_funs.R
\name{st}
\alias{st}
\title{Standardize whitespace in text}
\usage{
st(x)
}
\arguments{
\item{x}{Character vector to standardize}
}
\value{
Character vector with standardized whitespace
}
\description{
Standardize whitespace in text
}
