% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk.geocodes}
\alias{uk.geocodes}
\alias{geocodes.uk}
\title{UK geocode download URLs}
\description{
\code{uk.geocodes} and \code{geocodes.uk} contain various
details about URLs which can be used to download
boundaries. Two fields are compulsory: \code{url} and
\code{format}. But you may find \code{year} and other details
for the dataset.
}
\note{
The data can be (re-)generated using the script in the
\code{data-raw} directory.
}
\keyword{datasets}
