% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{gc_address}
\alias{gc_address}
\title{Parse Addresses}
\usage{
gc_address(data, address, city = NULL, zip = NULL, state = NULL, county = NULL)
}
\arguments{
\item{data}{A data frame}

\item{address}{The column containing the address. Can be a one-line address
or just the street address component (if \code{city}, \code{state}, etc. are
provided). Tidy-evaluated within \code{data}.}

\item{city}{A column containing city names. Tidy-evaluated within \code{data}.}

\item{zip}{A column containing ZIP codes. Tidy-evaluated within \code{data}.}

\item{state}{A column containing states. Tidy-evaluated within \code{data}.}

\item{county}{A column containing counties. Matched to an internal list and
passed on to the output. Tidy-evaluated within \code{data}.}
}
\value{
A data frame containing columns with the parsed address. These match
the Census format for street addresses.
}
\description{
Parse Addresses
}
\examples{
data(nc_addr)
gc_address(nc_addr)

}
