% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gc_import}
\alias{gc_import}
\title{Import geocode}
\usage{
gc_import(url, collection = "geocode", local = TRUE,
  index = "2dsphere", mongo_url = "mongodb://localhost:27017",
  silent = FALSE)
}
\arguments{
\item{url}{file to read or remote, flick \code{local} param.}

\item{collection}{set collection from parameter.}

\item{local}{assume \code{url} being local file.}

\item{index}{a MongoDB compatible index: either 2d or 2dsphere.}

\item{mongo_url}{custom mongourl to use with \code{collection},
default value is \code{'mongodb://localhost:27017'} used with \code{gc_setup}.}

\item{silent}{show messages, default is \code{FALSE}}
}
\value{
number of rows inserted
}
\description{
For now, support geojson file as that is what mongolite package tries to support.
Shape and others later which would also mean, keeping the package deps minimal.
For details see data-raw example of this package how geojson objects are imported.
}
