% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{geocodeR}
\alias{geocodeR}
\title{Input data}
\usage{
geocodeR()
}
\arguments{
\item{path}{Folder path of where the data will be output.}

\item{vars_regex}{Regex pattern to match variable names to input.}

\item{vars_vec}{A vector of variable names to input.}

\item{verbose}{Logical argument to be verbose.}
}
\description{
Output data to *.RDS objects.
}
